# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SentinelX – Autonomous AI Red Team vs Blue Team Cyber Range API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /simulation/start:
    post:
      operationId: startSimulation
      tags: [simulation]
      summary: Start a new simulation run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationConfig"
      responses:
        "200":
          description: Simulation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationStatus"
...
  /simulation/events:
    get:
      operationId: listSimulationEvents
      tags: [simulation]
      summary: List all simulation events (attack timeline)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: type
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimulationEvent"
...
components:
  schemas:
...
    SimulationStatus:
      type: object
      required: [id, state, phase, redTeamScore, blueTeamScore, startedAt]
      properties:
        id:
          type: integer
        state:
          type: string
          enum: [idle, running, paused, completed]
        phase:
          type: string
          enum: [reconnaissance, exploitation, persistence, lateral_movement, exfiltration, completed]
        redTeamScore:
          type: integer
        blueTeamScore:
          type: integer
        overallRisk:
          type: ["string", "null"]
          enum: [low, medium, high, critical, null]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
    SimulationConfig:
      type: object
      required: [scenario]
      properties:
        scenario:
          type: string
          enum: [apt_attack, ransomware, insider_threat, supply_chain, phishing_campaign]
        difficulty:
          type: string
          enum: [easy, medium, hard, expert]
          default: medium
        targetEnv:
          type: string
          default: "enterprise"
        autoAdvance:
          type: boolean
          default: true
    SimulationEvent:
      type: object
      required: [id, type, actor, description, severity, timestamp]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [attack, detection, response, compromise, containment, escalation]
        actor:
          type: string
          enum: [red_team, blue_team, commander, system]
        description:
          type: string
        severity:
          type: string
          enum: [info, low, medium, high, critical]
        assetId:
          type: ["integer", "null"]
        assetName:
          type: ["string", "null"]
        technique:
          type: ["string", "null"]
        mitreId:
          type: ["string", "null"]
        timestamp:
          type: string
...